refactor: extract types.ts from orchestrator god module [EE10.01]
Move all exported types and interfaces out of orchestrator.ts into a
dedicated types.ts module. Update all internal delivery modules to import
types from ./types instead of ./orchestrator. orchestrator.ts re-exports
all types from types.ts to preserve the public API.

Zero behavior change. bun test green.
