+++ tools/delivery/test/review.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test';
 
-import type { AiReviewFetcherResult } from '../orchestrator';
+import type { AiReviewFetcherResult } from '../types';
 import { buildStandaloneReviewStartedEvent } from '../pr-metadata';
 import {
   DEFAULT_REVIEW_POLLING_PROFILE,
